% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_mods.R
\name{comp_mods}
\alias{comp_mods}
\title{Model Performace Comparison}
\usage{
comp_mods(mod1, ..., newdata = NULL)
}
\arguments{
\item{mod1}{an object containing results returned by lm function}

\item{...}{additional lm model object}

\item{newdata}{newdata than used for training the model. By default NULL and the function evaluated model performance based on training data}
}
\value{
a dataframe with each row representing an input model and three columns of model performance evaluation criteria (RMSE, R2 and MAE)
}
\description{
This function takes in one or more linear regression models and compares the RMSE, R2 and MAE of those models with or without a new input data.
}
\examples{
train <- mtcars[1:20,]
test <- mtcars[21:30,]

mod1 <- lm(mpg~wt, train)
mod2 <- lm(mpg~cyl, train)
mod3 <- lm(mpg~wt+cyl, train)
mod4 <- lm(mpg~carb, train)

comp_mods(mod1, mod2, mod3, mod4, newdata = test)

}
