% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_significant.R
\name{lm_significant}
\alias{lm_significant}
\title{Model Optimization to significant predictors}
\usage{
lm_significant(
  data,
  res,
  preds = NULL,
  p = 0.01,
  verbose = TRUE,
  all = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame object containing the variables to be used as response and predictors in the model.}

\item{res}{a character vector of length 1 that matches the name of the response variable column in data. Response variable in the data must be of numeric type.}

\item{preds}{a character vector of predictor variables in the data. When not specified function will take all variables in data other than response variable as input predictors to start with. A default of NULL is given to this argument to provide the flexibility of using either user defined predictor variables or all but response variable as predictors from the data.}

\item{p}{a numeric value that denotes the threshold for selecting the predictors based on their statistical significance in building the model. Default p-value threshold is 0.01.}

\item{verbose}{a logical value denoting whether or not to print progress messages as the function is being run. Default is TRUE}

\item{all}{if TRUE, the function will return a list with two lm model objects. The first one is the original call with all the input predictors and the second is the model with only the significant predictors. The default is FALSE which returns a single lm object with only the significant predictors}

\item{...}{additional arguments to be passed to the inner lm() function calls. Refer to the documentation of lm() for more details on those arguments}
}
\value{
The function will return a list with two lm model objects is all = TRUE. The first one is the original call with all the input predictors and the second is the model with only the significant predictors.

If all = FALSE the function returns a single lm object with only the significant predictors
}
\description{
lm_significant is used to optimize a multiple linear regression model to only those predictor variables that are statistically significant for building that model. The function currently only supports additive model linear regressions.
}
\examples{
#cancer_sample data from datateachr package

library(datateachr)

sig_mod1 <- lm_significant(cancer_sample[,-2], res = "radius_mean")

#Both the original model and optimized model

sig_mod2 <- lm_significant(mtcars, res = "mpg", all = TRUE)

}
